refactor(agent-tooling): dedupe quiet-build capture-and-replay logic
scripts/build-agent.mjs, scripts/bundle-sizes.ts's self-build, and
scripts/prepare.mjs's quiet build each hand-rolled the same "capture
stdout/stderr instead of inheriting, replay both plus the exit code
only on failure" block. Extract it into scripts/lib/quiet-spawn.mjs
(with a sibling .d.mts, matching the lint-package-layers.mjs /
freeze-fold-stage0-compatibility.mjs precedent for a .ts consumer of a
plain .mjs module) and import it from all three call sites.

The extraction also fixes a latent correctness gap flagged by review:
spawnSync's default maxBuffer is 1 MiB per stream, and a verbose-but-
successful build exceeding it gets its child process killed (status:
null), which every one of these call sites would have then reported
as a spurious failure. The shared helper passes an explicit 64 MiB
maxBuffer, closing the gap in prepare.mjs's pre-existing instance too.
