+++ scripts/lib/quiet-spawn.d.mts
@@ -1,0 +1,4 @@
+// Type declarations for scripts/lib/quiet-spawn.mjs. Consumed by
+// scripts/bundle-sizes.ts, which tsgo type-checks under `pnpm verify`.
+
+export function spawnQuiet(cmd: string, args: readonly string[]): void;
